add authentication

+++ backend/controllers/goalsController.js
@@ -2,16 +2,18 @@ const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModel');
 
 // @desc Get goals
-// @route GET /api/goals
+// @route GET /api/v1/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({
+        user: req.user.id
+    });
 
     res.status(200).json({goals});
 })
 
 // @desc Get goal
-// @route GET /api/goals/:id
+// @route GET /api/v1/goals/:id
 // @access Private
 const getGoal = asyncHandler(async (req, res) => {
     const goal = await Goal.findById(req.params.id);
@@ -21,11 +23,16 @@ const getGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal not found.');
     }
 
+    if (goal.user.toString() !== req.user.id) {
+        res.status(404);
+        throw new Error("Goal not found.")
+    }
+
     return res.status(200).json({goal})
 })
 
-// @desc Set goal
-// @route POST /api/goals/
+// @desc Create goal
+// @route POST /api/v1/goals/
 // @access Private
 const createGoal = asyncHandler(async (req, res) => {
     if (!req.body.text) {
@@ -34,6 +41,7 @@ const createGoal = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
+        user: req.user.id,
         text: req.body.text
     });
 
@@ -41,7 +49,7 @@ const createGoal = asyncHandler(async (req, res) => {
 })
 
 // @desc Delete goal
-// @route DELETE /api/goals/:id
+// @route DELETE /api/v1/goals/:id
 // @access Private
 const deleteGoal = asyncHandler(async (req, res) => {
     const id = req.params.id;
@@ -53,6 +61,11 @@ const deleteGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal not found');
     }
 
+    if (goal.user.toString() !== req.user.id) {
+        res.status(404);
+        throw new Error('Goal not found.');
+    }
+
     await Goal.findByIdAndDelete(id)
 
     res.status(200).send({
@@ -61,7 +74,7 @@ const deleteGoal = asyncHandler(async (req, res) => {
 })
 
 // @desc Update goal
-// @route PUT /api/goals/:id
+// @route PUT /api/v1/goals/:id
 // @access Private
 const updateGoal = asyncHandler(async (req, res) => {
     if (! req.body.text) {
@@ -74,6 +87,11 @@ const updateGoal = asyncHandler(async (req, res) => {
     const goal = await Goal.findById(id);
 
     if (! goal) {
+        res.status(404);
+        throw new Error("Goal not found.");
+    }
+
+    if (goal.user.toString() !== req.user.id) {
         res.status(404);
         throw new Error("Goal not found.");
     }
